requirements.txt
# Specify Python version for deployment environments like Streamlit Cloud
# python==3.11

# Core Web & App Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
streamlit-authenticator==0.2.3

# Data Science & Machine Learning
pandas==2.1.3
numpy==1.26.0  # Updated for compatibility with pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
plotly==5.17.0

# AI, NLP & Vector Search
google-generativeai==0.3.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4  # Pinned for compatibility with sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Utility & Data Connectors
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8  # 'fitz' is the import name for this library